+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
  - Add interface `DataWareActionFromBuilder` to allow actions define default form values
+ - Add update entity action for dca using `useRawRequestData` flag
 
 ## [2.2.4] - 2020-10-08